refactor(recovery): extract format state capture and facade wiring
